requirements.txt
fastapi
uvicorn
passlib
bcrypt==4.0.1
SQLAlchemy
python-jose[cryptography]
psycopg2-binary
python-dotenv
